Moved message area into its own component
