+++ web/src/App.jsx
@@ -1,22 +1,12 @@
-import { useState, useEffect, useRef } from "react";
+import { useState } from "react";
 
 import MessageForm from "./components/MessageForm";
+import MessageArea from "./components/MessageArea";
 import * as Constants from "./utils/constants";
 
 function App() {
-
   const [chatHistory, setChatHistory] = useState([Constants.INTRO_MESSAGE]);
   const [currentChat, setCurrentChat] = useState("");
-  const historyAreaRef = useRef(null);
-
-  useEffect(() => {
-    historyAreaRef.current.scrollTop = historyAreaRef.current.scrollHeight;
-  }, [chatHistory, currentChat]);
-
-  eel.expose(js_app__updateCurrentChat);
-  function js_app__updateCurrentChat(chunk) {
-    setCurrentChat((prevState) => prevState + chunk);
-  }
 
   const onSubmitForm = (textValue) => {
     if (currentChat === "") {
@@ -24,9 +14,7 @@ function App() {
     } else {
       setChatHistory([...chatHistory, currentChat, `\n${Constants.USER_ROLE_NAME}: ${textValue}`]);
     }
-
     setCurrentChat(`${Constants.ASSISTANT_ROLE_NAME}: `);
-
     eel.py_main__send_chat(textValue)();
   };
 
@@ -34,24 +22,24 @@ function App() {
     setCurrentChat("");
     setChatHistory([Constants.INTRO_MESSAGE]);
     eel.py_main__clear_chats()();
+  };
+
+  eel.expose(js_app__updateCurrentChat);
+  function js_app__updateCurrentChat(chunk) {
+    setCurrentChat((prevState) => prevState + chunk);
   }
 
   return (
     <div className="mg-container">
-
-      <div className="mg-row">
-
-        <div className="mg-col mg-x12 history">
-          <pre><code ref={historyAreaRef}>
-            {chatHistory.map((chat) => chat + "\n")}
-            {currentChat && (currentChat + "\n") }
-          </code></pre>
-        </div>
+      <div className="mg-row message-area">
+        <MessageArea chatHistory={chatHistory} currentChat={currentChat} />
       </div>
-
-      <MessageForm onSubmitForm={onSubmitForm} onClearForm={onClearForm} className="mg-row message" />
+      <div className="mg-row message-form">
+        <MessageForm onSubmitForm={onSubmitForm} onClearForm={onClearForm} />
+      </div>
     </div>
   );
 }
 
 export default App;
+
